requirements/development.txt
# Development requirements
-r base.txt

# Development tools
django-debug-toolbar==6.1.0
django-extensions==4.1

# Testing
pytest==9.0.2
pytest-django==4.11.1
pytest-cov==7.0.0
factory-boy==3.3.3

# Code quality - Updated Nov 1, 2025 (Phase 16.2)
black==25.12.0
flake8==7.3.0
flake8-docstrings==1.7.0
isort==7.0.0
mypy==1.19.1
mypy-extensions==1.1.0
bandit==1.9.2
django-stubs==5.2.8
django-stubs-ext==5.2.8
types-PyYAML==6.0.12.20250915
types-requests==2.32.4.20260107

# Performance monitoring
django-silk==5.4.3
